+++ actix-rt/src/lib.rs
@@ -61,6 +61,7 @@ mod arbiter;
 mod runtime;
 mod system;
 
+#[deprecated(since = "2.11.0", note = "Prefer `std::pin::pin!`.")]
 pub use tokio::pin;
 use tokio::task::JoinHandle;
 
@@ -87,10 +88,11 @@ pub mod net {
     use std::{
         future::Future,
         io,
+        pin::pin,
         task::{Context, Poll},
     };
 
-    use tokio::io::{AsyncRead, AsyncWrite, Interest};
+    use tokio::io::{AsyncRead, AsyncWrite, BufReader, Interest};
     #[cfg(unix)]
     pub use tokio::net::{UnixDatagram, UnixListener, UnixStream};
     pub use tokio::{
@@ -115,14 +117,12 @@ pub mod net {
     impl ActixStream for TcpStream {
         fn poll_read_ready(&self, cx: &mut Context<'_>) -> Poll<io::Result<Ready>> {
             let ready = self.ready(Interest::READABLE);
-            tokio::pin!(ready);
-            ready.poll(cx)
+            pin!(ready).poll(cx)
         }
 
         fn poll_write_ready(&self, cx: &mut Context<'_>) -> Poll<io::Result<Ready>> {
             let ready = self.ready(Interest::WRITABLE);
-            tokio::pin!(ready);
-            ready.poll(cx)
+            pin!(ready).poll(cx)
         }
     }
 
@@ -130,14 +130,12 @@ pub mod net {
     impl ActixStream for UnixStream {
         fn poll_read_ready(&self, cx: &mut Context<'_>) -> Poll<io::Result<Ready>> {
             let ready = self.ready(Interest::READABLE);
-            tokio::pin!(ready);
-            ready.poll(cx)
+            pin!(ready).poll(cx)
         }
 
         fn poll_write_ready(&self, cx: &mut Context<'_>) -> Poll<io::Result<Ready>> {
             let ready = self.ready(Interest::WRITABLE);
-            tokio::pin!(ready);
-            ready.poll(cx)
+            pin!(ready).poll(cx)
         }
     }
 
@@ -148,6 +146,16 @@ pub mod net {
 
         fn poll_write_ready(&self, cx: &mut Context<'_>) -> Poll<io::Result<Ready>> {
             (**self).poll_write_ready(cx)
+        }
+    }
+
+    impl<Io: ActixStream> ActixStream for BufReader<Io> {
+        fn poll_read_ready(&self, cx: &mut Context<'_>) -> Poll<io::Result<Ready>> {
+            self.get_ref().poll_read_ready(cx)
+        }
+
+        fn poll_write_ready(&self, cx: &mut Context<'_>) -> Poll<io::Result<Ready>> {
+            self.get_ref().poll_write_ready(cx)
         }
     }
 }